Move files out of opm/core/simulator.
